refactor: split interface into two files, update docs, add integration tests
