+++ crates/miden-lib/src/transaction/mod.rs
@@ -34,8 +34,8 @@ pub use errors::{TransactionEventError, TransactionKernelError, TransactionTrace
 
 mod procedures;
 
-mod script_builder;
-pub use script_builder::{TransactionScriptBuilder, TransactionScriptBuilderError};
+// mod script_builder;
+// pub use script_builder::{TransactionScriptBuilder, TransactionScriptBuilderError};
 
 // CONSTANTS
 // ================================================================================================
