+++ crates/miden-lib/src/transaction/script_builder/mod.rs
@@ -50,36 +50,16 @@ impl TransactionScriptBuilder {
         }
     }
 
-    /// Builds a transaction script which sends the specified notes with the corresponding
-    /// authentication.
-    pub fn build_send_notes_script(
-        &self,
-        output_notes: &[PartialNote],
-    ) -> Result<TransactionScript, TransactionScriptBuilderError> {
-        let send_note_procedure = if self
-            .account_interface
-            .components()
-            .contains(&AccountComponentInterface::BasicFungibleFaucet)
-        {
-            AccountComponentInterface::BasicFungibleFaucet
-                .send_note_procedure(*self.account_interface.id(), output_notes)?
-        } else if self
-            .account_interface
-            .components()
-            .contains(&AccountComponentInterface::BasicWallet)
-        {
-            AccountComponentInterface::BasicWallet
-                .send_note_procedure(*self.account_interface.id(), output_notes)?
-        } else {
-            return Err(TransactionScriptBuilderError::UnsupportedAccountInterface);
-        };
-
-        self.build_script_with_sections(&[send_note_procedure])
-    }
-
     /// Builds a simple authentication script for the transaction that doesn't send any notes.
-    pub fn build_auth_script(&self) -> Result<TransactionScript, TransactionScriptBuilderError> {
-        self.build_script_with_sections(&[])
+    pub fn build_auth_script(
+        &self,
+        in_debug_mode: bool,
+    ) -> Result<TransactionScript, TransactionScriptBuilderError> {
+        let auth_script_source = format!("begin\n{}\nend", self.script_authentication());
+        let assembler = TransactionKernel::assembler().with_debug_mode(in_debug_mode);
+
+        TransactionScript::compile(auth_script_source, [], assembler)
+            .map_err(TransactionScriptBuilderError::InvalidTransactionScript)
     }
 
     /// Builds a transaction script with the specified sections.
@@ -87,19 +67,25 @@ impl TransactionScriptBuilder {
     /// The `sections` parameter is a slice of strings, where each string represents a distinct
     /// part of the script body. The script authentication and expiration sections are
     /// automatically added to the script.
-    fn build_script_with_sections(
+    pub fn build_send_notes_script(
         &self,
-        sections: &[String],
+        output_notes: &[PartialNote],
+        expiration_delta: Option<u16>,
+        in_debug_mode: bool,
     ) -> Result<TransactionScript, TransactionScriptBuilderError> {
+        let note_creation_source = self.note_creation(output_notes)?;
+
         let script = format!(
-            "begin{}\n{}\n{}\nend",
+            "begin\n{}\n{}\n{}\nend",
             self.script_expiration(),
-            sections.join(" "),
+            note_creation_source,
             self.script_authentication()
         );
 
-        let assembler = TransactionKernel::assembler().with_debug_mode(self.in_debug_mode);
-        let tx_script = TransactionScript::compile(script, vec![], assembler)
+        std::println!("{script}");
+
+        let assembler = TransactionKernel::assembler().with_debug_mode(in_debug_mode);
+        let tx_script = TransactionScript::compile(script, [], assembler)
             .map_err(TransactionScriptBuilderError::InvalidTransactionScript)?;
 
         Ok(tx_script)
@@ -111,17 +97,65 @@ impl TransactionScriptBuilder {
         self.account_interface.auth().iter().for_each(|auth_scheme| match auth_scheme {
             &AuthScheme::RpoFalcon512 { pub_key: _ } => {
                 auth_script
-                    .push_str("\n\tcall.::miden::contracts::auth::basic::auth_tx_rpo_falcon512");
+                    .push_str("call.::miden::contracts::auth::basic::auth_tx_rpo_falcon512\n");
             },
         });
 
         auth_script
     }
 
+    /// Builds a transaction script which sends the specified notes with the corresponding
+    /// authentication.
+    ///
+    ///
+    /// Generates a `send_note` procedure. The resulting code differs depending on the used
+    /// interface.
+    ///
+    /// The body consists of two sections:
+    /// - Pushing the note information on the stack.
+    /// - Creating a note:
+    ///   - For basic fungible faucet: pushing the amount of assets and distributing them.
+    ///   - For basic wallet: creating a note, pushing the assets on the stack and moving them to
+    ///     the created note.
+    ///
+    /// # Errors:
+    /// Returns an error if:
+    /// - If the interface does not support the generation of the standard `send_note` procedure.
+    /// - If the sender of the note isn't the account for which the script is being built.
+    /// - If the note created by the faucet doesn't contain exactly one asset.
+    /// - If a faucet tries to distribute an asset with a different faucet ID.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// ```
+    fn note_creation(
+        &self,
+        output_notes: &[PartialNote],
+    ) -> Result<String, TransactionScriptBuilderError> {
+        if self
+            .account_interface
+            .components()
+            .contains(&AccountComponentInterface::BasicFungibleFaucet)
+        {
+            AccountComponentInterface::BasicFungibleFaucet
+                .send_note_procedure(*self.account_interface.id(), output_notes)
+        } else if self
+            .account_interface
+            .components()
+            .contains(&AccountComponentInterface::BasicWallet)
+        {
+            AccountComponentInterface::BasicWallet
+                .send_note_procedure(*self.account_interface.id(), output_notes)
+        } else {
+            return Err(TransactionScriptBuilderError::UnsupportedAccountInterface);
+        }
+    }
+
     /// Returns a string with the expiration delta update procedure call for the script.
     fn script_expiration(&self) -> String {
         if let Some(expiration_delta) = self.expiration_delta {
-            format!("\n\tpush.{expiration_delta} exec.::miden::tx::update_expiration_block_delta")
+            format!("push.{expiration_delta} exec.::miden::tx::update_expiration_block_delta\n")
         } else {
             String::new()
         }
@@ -132,18 +166,17 @@ impl TransactionScriptBuilder {
 // ============================================================================================
 
 /// Errors related to building a transaction script.
-#[derive(Debug, Error)]
-pub enum TransactionScriptBuilderError {
-    #[error("note asset is not issued by this faucet: {0}")]
-    IssuanceFaucetMismatch(AccountIdPrefix),
-    #[error("note created by the faucet doesn't contain exactly one asset")]
-    FaucetNoteWithoutAsset,
-    #[error("invalid transaction script")]
-    InvalidTransactionScript(#[source] TransactionScriptError),
-    #[error("invalid sender account: {0}")]
-    InvalidSenderAccount(AccountId),
-    #[error("{0} interface does not support the generation of the standard send_note script")]
-    UnsupportedInterface(AccountComponentInterface),
-    #[error("account does not contain the basic fungible faucet or basic wallet interfaces which are needed to support the send_note script generation")]
-    UnsupportedAccountInterface,
-}
+// #[derive(Debug, Error)]
+// pub enum TransactionScriptBuilderError {
+//     #[error("note asset is not issued by this faucet: {0}")]
+//     IssuanceFaucetMismatch(AccountIdPrefix),
+
+//     #[error("invalid transaction script")]
+//     InvalidTransactionScript(#[source] TransactionScriptError),
+//     #[error("invalid sender account: {0}")]
+//     InvalidSenderAccount(AccountId),
+//     #[error("{0} interface does not support the generation of the standard send_note script")]
+//     UnsupportedInterface(AccountComponentInterface),
+//     #[error("account does not contain the basic fungible faucet or basic wallet interfaces which are needed to support the send_note script generation")]
+//     UnsupportedAccountInterface,
+// }
