+++ crates/miden-objects/src/lib.rs
@@ -71,7 +71,7 @@ pub mod utils {
     /// let word = Word::from([Felt::new(1), Felt::new(2), Felt::new(3), Felt::new(4)]);
     /// assert_eq!(word_to_felts_string(&word), "1.2.3.4");
     /// ```
-    pub fn word_to_felts_string(word: &Word) -> String {
+    pub fn word_to_masm_push_string(word: &Word) -> String {
         format!("{}.{}.{}.{}", word[0], word[1], word[2], word[3])
     }
