+++ crates/miden-tx/src/tests/kernel_tests/test_asset.rs
@@ -10,7 +10,7 @@ use miden_objects::{
 };
 use vm_processor::ProcessState;
 
-use super::{word_to_felts_string, Felt, Hasher, Word, ONE};
+use super::{word_to_masm_push_string, Felt, Hasher, Word, ONE};
 use crate::testing::TransactionContextBuilder;
 
 #[test]
@@ -83,7 +83,7 @@ fn test_create_non_fungible_asset_succeeds() {
         end
         ",
         non_fungible_asset_data_hash =
-            word_to_felts_string(&Hasher::hash(&NON_FUNGIBLE_ASSET_DATA)),
+            word_to_masm_push_string(&Hasher::hash(&NON_FUNGIBLE_ASSET_DATA)),
     );
 
     let process = &tx_context.execute_code(&code).unwrap();
@@ -116,7 +116,7 @@ fn test_validate_non_fungible_asset() {
             swapw dropw
         end
         ",
-        asset = word_to_felts_string(&encoded)
+        asset = word_to_masm_push_string(&encoded)
     );
 
     let process = &tx_context.execute_code(&code).unwrap();
