+++ crates/miden-tx/tests/integration/scripts/mod.rs
@@ -1,4 +1,5 @@
 mod faucet;
 mod p2id;
 mod p2idr;
+mod send_note;
 mod swap;
